+++ .env.example
@@ -3,5 +3,12 @@ DB_USER=
 DB_PWD=
 DB_NAME=
 JWT_SECRET=
+SECRET_KEY=
 PORT=
 MATCH_BET_AMOUNT=
+
+# AWS S3 Configuration
+AWS_ACCESS_KEY_ID=
+AWS_SECRET_ACCESS_KEY=
+AWS_BUCKET_NAME=
+AWS_REGION=
