+++ config.js
@@ -1,12 +1,12 @@
 module.exports = {
     database: {
-      host: 'saitejareddy-chappetasaitejareddy-75bf.i.aivencloud.com',
-      name: 'tolaram',
-      user: 'avnadmin',
-      password: 'AVNS_D_8TtIDpt1WUkZ9a3Qy',
-      port: 25794
+      host: '',
+      name: '',
+      user: '',
+      password: '',
+      port: ''
     },
-    server_port : 3000,
-    jwtSecret: 'secret'
+    server_port : '',
+    jwtSecret: ''
   };
