+++ controllers/players.js
@@ -13,76 +13,6 @@ function validateString(string) {
   return true;
 }
 
-router.patch("/:playerId", verifyRole("admin"), async (req, res) => {
-    const { playerId } = req.params;
-    const { name, imageKey, role  } = req.body;
-  
-    if (name && !validateName(name)) {
-      return res.status(400).json({ message: "Invalid player name" });
-    }
-  
-    if (imageKey && !validateString(imageKey)) {
-      return res.status(400).json({ message: "Invalid key" });
-    }
-    if (role && !['all-rounder', 'batsman', 'bowler', 'wicket-keeper'].includes(role)) {
-        return res.status(400).json({ message: "Invalid role for player" });
-      }
-
-  
-    try {
-      // Check if the team exists
-      const [playerRows] = await pool.execute(
-        "SELECT * FROM players WHERE id = ?",
-        [playerId]
-      );
-  
-      if (playerRows.length === 0) {
-        return res.status(404).json({ message: "Player not found" });
-      }
-  
-      // Update fields conditionally
-      const updateFields = [];
-      const updateValues = [];
-  
-      if (name) {
-        updateFields.push("name = ?");
-        updateValues.push(name);
-      }
-  
-      if (imageKey) {
-        updateFields.push("player_logo = ?");
-        updateValues.push(imageKey);
-      }
-  
-      if (role !== undefined) {
-        updateFields.push("player_role = ?");
-        updateValues.push(role);
-      }
-  
-      if (updateFields.length === 0) {
-        return res.status(400).json({ message: "No fields to update" });
-      }
-  
-      updateValues.push(playerId);
-  
-      const [updateResult] = await pool.execute(
-        `UPDATE players SET ${updateFields.join(", ")} WHERE id = ?`,
-        updateValues
-      );
-  
-      if (updateResult.affectedRows === 0) {
-        return res.status(400).json({ message: "Update failed" });
-      }
-  
-      res.json({
-        message: "Player updated successfully"
-      });
-    } catch (error) {
-      console.error("Error executing query", error);
-      res.status(500).json({ message: "Internal server error" });
-    }
-  });
-
   router.get("/", verifyToken, async (req, res) => {
     try {
       const [players] = await pool.execute("SELECT p.* , t.team_name , t.team_logo FROM players p LEFT JOIN teams t ON p.team_id = t.id");
