+++ controllers/teams.js
@@ -13,116 +13,6 @@ function validateString(string) {
   return true;
 }
 
-router.post("/", verifyRole("admin"), async (req, res) => {
-  const { name, imageKey } = req.body;
-
-  if (!validateName(name)) {
-    res.status(400).json({ message: "Invalid team name" });
-    return;
-  }
-
-  if (!validateString(imageKey)) {
-    res.status(400).json({ message: "Invalid key" });
-    return;
-  }
-
-  try {
-    const [teamRows] = await pool.execute(
-      "SELECT * FROM teams WHERE team_name = ?",
-      [name]
-    );
-
-    if (teamRows.length > 0) {
-      res.status(400).json({ message: "Team already exists" });
-      return;
-    }
-
-    const [insertResult] = await pool.execute(
-      "INSERT INTO teams (team_name, team_logo) VALUES (?, ?)",
-      [name, imageKey]
-    );
-
-    res.json({
-      message: "Team created successfully",
-      teamId: insertResult.insertId,
-    });
-  } catch (error) {
-    console.error("Error executing query", error);
-    res.status(500).json({ message: "Internal server error" });
-  }
-});
-
-router.patch("/:teamId", verifyRole("admin"), async (req, res) => {
-    const { teamId } = req.params;
-    const { name, imageKey, status } = req.body;
-  
-    if (name && !validateName(name)) {
-      return res.status(400).json({ message: "Invalid team name" });
-    }
-  
-    if (imageKey && !validateString(imageKey)) {
-      return res.status(400).json({ message: "Invalid key" });
-    }
-  
-    if (status && !['0', '1'].includes(status)) {
-      return res.status(400).json({ message: "Invalid operation" });
-    }
-  
-    try {
-      // Check if the team exists
-      const [teamRows] = await pool.execute(
-        "SELECT * FROM teams WHERE id = ?",
-        [teamId]
-      );
-  
-      if (teamRows.length === 0) {
-        return res.status(404).json({ message: "Team not found" });
-      }
-  
-      // Update fields conditionally
-      const updateFields = [];
-      const updateValues = [];
-  
-      if (name) {
-        updateFields.push("team_name = ?");
-        updateValues.push(name);
-      }
-  
-      if (imageKey) {
-        updateFields.push("team_logo = ?");
-        updateValues.push(imageKey);
-      }
-  
-      if (status !== undefined) {
-        updateFields.push("status = ?");
-        updateValues.push(status);
-      }
-  
-      if (updateFields.length === 0) {
-        return res.status(400).json({ message: "No fields to update" });
-      }
-  
-      // Add teamId to the end for where clause
-      updateValues.push(teamId);
-  
-      // Perform the update query
-      const [updateResult] = await pool.execute(
-        `UPDATE teams SET ${updateFields.join(", ")} WHERE id = ?`,
-        updateValues
-      );
-  
-      if (updateResult.affectedRows === 0) {
-        return res.status(400).json({ message: "Update failed" });
-      }
-  
-      res.json({
-        message: "Team updated successfully"
-      });
-    } catch (error) {
-      console.error("Error executing query", error);
-      res.status(500).json({ message: "Internal server error" });
-    }
-  });
 
   router.get("/", verifyToken, async (req, res) => {
     try {
@@ -187,61 +77,5 @@ router.patch("/:teamId", verifyRole("admin"), async (req, res) => {
     }
   });
 
-  router.post("/:teamId/addPlayer", verifyRole("admin"), async (req, res) => {
-    const { teamId } = req.params;
-    const { name, imageKey , role} = req.body;
-  
-    if (!validateName(name)) {
-      return res.status(400).json({ message: "Invalid player name" });
-    }
-  
-    if (!validateString(imageKey)) {
-      return res.status(400).json({ message: "Invalid key" });
-    }
-
-    if (!role || !['all-rounder', 'batsman', 'bowler', 'wicket-keeper'].includes(role)) {
-        return res.status(400).json({ message: "Invalid role for player" });
-      }
-
-  
-    try {
-      // Check if the team exists
-      const [teamRows] = await pool.execute(
-        "SELECT * FROM teams WHERE id = ?",
-        [teamId]
-      );
-  
-      if (teamRows.length === 0) {
-        return res.status(404).json({ message: "Team not found" });
-      }
-
-      const [playerRows] = await pool.execute(
-        "SELECT * FROM players WHERE name = ?",
-        [name]
-      );
-  
-      if (playerRows.length > 0) {
-        return res.status(400).json({ message: "Player with the name already exists" });
-      }
-  
-      const [insertResult] = await pool.execute(
-        "INSERT INTO players (name, player_logo, team_id, player_role) VALUES (?, ?, ?, ?)",
-        [name, imageKey, teamId, role]
-      );
-  
-      res.json({
-        message: "Player added successfully",
-        player: {
-          id: insertResult.insertId,
-          name,
-          imageKey,
-          teamId,
-        },
-      });
-    } catch (error) {
-      console.error("Error executing query", error);
-      res.status(500).json({ message: "Internal server error" });
-    }
-  });
 
 module.exports = router;
