+++ server.js
@@ -11,6 +11,7 @@ const moment = require("moment");
 require("moment-timezone");
 moment.tz.setDefault("Africa/Lagos");
 
+
 const loginRouter = require("./auth/login");
 const signupRouter = require("./auth/signup");
 
@@ -21,7 +22,10 @@ const betHistory = require("./controllers/betHistory");
 const teamsRouter = require("./controllers/teams");
 const mediaRouter = require("./controllers/media");
 const playersRouter = require("./controllers/players");
-const userRouter = require("./controllers/user");
+const dream11 = require("./controllers/dream11");
+const adminRouter = require("./controllers/admin");
+const commentsRouter = require("./controllers/comments");
+const { verifyToken, verifyRole } = require("./middleware/middleware");
 
 app.use("/login", loginRouter);
 app.use("/signup", signupRouter);
@@ -32,8 +36,9 @@ app.use("/history", betHistory);
 app.use("/teams", teamsRouter);
 app.use("/media", mediaRouter);
 app.use("/players", playersRouter);
-app.use('/user', userRouter);
-
+app.use('/dream11', dream11);
+app.use('/comments', commentsRouter);
+app.use('/admin', verifyToken, verifyRole('admin'), adminRouter);
 // Start the server
 const port = process.env.PORT;
 app.listen(port, () => {
